refactor-hooks-part-deux * Rename use-operators directory to use-operator-ids
