+++ src/logic/use-operator-ids/useOperatorIds.mobility.graphql
